[CE] fix: role_id migration failure

+++ server/db/migrate/20240515072318_add_role_id_to_workspace_users.rb
@@ -1,6 +1,10 @@
 class AddRoleIdToWorkspaceUsers < ActiveRecord::Migration[7.1]
   def change
-    remove_column :workspace_users, :role
-    add_reference :workspace_users, :role, null: false, foreign_key: true, default: Role.find_by(role_name: "Admin").id
+    if column_exists?(:workspace_users, :role)
+      remove_column :workspace_users, :role
+    end
+    unless column_exists?(:workspace_users, :role_id)
+      add_reference :workspace_users, :role, null: false, foreign_key: true, default: nil
+    end
   end
 end
